# contract/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateItemInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/CreateItemInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "category_id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "description": "Image URIs returned by POST /api/merchant/uploads",
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "price_minor": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "description",
          "price_minor",
          "tags"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ItemDTO": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "price_minor": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "vendor_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "vendor_id",
          "name",
          "description",
          "price_minor",
          "tags",
          "status"
        ],
        "type": "object"
      },
      "ItemOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ItemOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "item": {
            "$ref": "#/components/schemas/ItemDTO"
          }
        },
        "required": [
          "item"
        ],
        "type": "object"
      },
...
      "ListItemsOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ListItemsOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MerchantItemDTO"
            },
            "type": [
              "array",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "MerchantItemDTO": {
        "additionalProperties": false,
        "properties": {
          "category_id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "last_used": {
            "description": "Most recent supply date (YYYY-MM-DD), null if never scheduled",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "price_minor": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "total_sold": {
            "description": "Cumulative quantity sold across picked-up orders",
            "format": "int64",
            "type": "integer"
          },
          "vendor_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "vendor_id",
          "name",
          "description",
          "price_minor",
          "tags",
          "status",
          "last_used",
          "total_sold"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "T-Bite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/merchant/menu-items": {
      "get": {
        "operationId": "listMerchantMenuItems",
        "parameters": [
          {
            "description": "Include archived items in the result",
            "explode": false,
            "in": "query",
            "name": "include_archived",
            "schema": {
              "description": "Include archived items in the result",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemsOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the merchant's menu items",
        "tags": [
          "merchant",
          "menu"
        ]
      },
      "post": {
        "operationId": "createMerchantMenuItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft menu item",
        "tags": [
          "merchant",
          "menu"
        ]
      }
    },
...
    "/api/merchant/menu-items/{id}/archive": {
      "post": {
        "operationId": "archiveMerchantMenuItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive a menu item",
        "tags": [
          "merchant",
          "menu"
        ]
      }
    },
    "/api/merchant/menu-items/{id}/copy": {
      "post": {
        "operationId": "copyMerchantMenuItem",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Duplicate a menu item into a fresh draft",
        "tags": [
          "merchant",
          "menu"
        ]
      }
    },
...
  }
}